requirements.txt
requests
pandas
python-dateutil
playwright
beautifulsoup4
python-dotenv
openai
